% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine.similarity.cutoff.R
\name{cosine.similarity.cutoff}
\alias{cosine.similarity.cutoff}
\title{Calculate an outlier cutoff using cosine similarity}
\usage{
cosine.similarity.cutoff(
  quality.scores,
  no.simulations,
  distribution = c("lnorm", "weibull", "norm", "gamma", "exp", "cauchy", "logis"),
  trim.factor = 0.05,
  alpha.significant = 0.05
)
}
\arguments{
\item{quality.scores}{A dataframe with columns 'Sum' (of scores) and 'Sample', i.e. the output of accumulate.zscores}

\item{no.simulations}{The number of datasets to simulate}

\item{distribution}{A distribution to test, will default to 'lnorm'}

\item{trim.factor}{What fraction of values of each to trim to get parameters without using extremes}

\item{alpha.significant}{Alpha value for significance}
}
\description{
This function takes quality.scores, trims it and fits it to the distribution given.
It then simulates as many datasets as stated by no.simulations, and computes the cosine
similarity of each dataset against theoretical distribution. It uses what would correspond
to a significant value to then calculate what observed value this would correspond to.
The function supports the following distributions:
\itemize{
\item 'weibull'
\item 'norm'
\item 'gamma'
\item 'exp'
\item 'lnorm'
\item 'cauchy'
\item 'logis'
}
}
