% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.outlier.detection.R
\name{cosine.similarity.iterative.testing}
\alias{cosine.similarity.iterative.testing}
\title{Tests the accumulated quality scores for outliers using cosine similarity}
\usage{
cosine.similarity.iterative.testing(
  quality.scores,
  no.simulations,
  distribution,
  trim.factor = 0.05,
  alpha.significant = 0.05
)
}
\arguments{
\item{quality.scores}{The output of accumulate.zscores}

\item{no.simulations}{The number of datasets to simulate}

\item{distribution}{A distribution to test}

\item{trim.factor}{What fraction of values of each to trim to get parameters without using extremes}

\item{alpha.significant}{Alpha value for significance}
}
\description{
This function takes quality.scores, trims it and fits it to the distribution given.
It then iterratively tests the largest datapoint compared a null distribution of size
no.simulations. If the largest datapoint has a significant p-value it tests the 2nd largest
one and so on. The function supports the following distributions:
\itemize{
\item 'weibull'
\item 'norm'
\item 'gamma'
\item 'exp'
\item 'lnorm'
\item 'cauchy'
\item 'logis'
}
}
