% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example.qc.dataframe}
\alias{example.qc.dataframe}
\title{QC metrics across 100 samples}
\format{
A data frame containing QC data; columns represent QC metrics and rows represent samples
}
\usage{
data(example.qc.dataframe)
}
\description{
QC metrics across 100 samples
}
\examples{
\donttest{
data(example.qc.dataframe)
zscores.from.metrics(
  qc.data = example.qc.dataframe
  );
}
}
\keyword{datasets}
