% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.standard.plots.R
\name{get.qc.barplot}
\alias{get.qc.barplot}
\title{Generates the standard barplot of scores for each sample}
\usage{
get.qc.barplot(
  barplot.data,
  filename = NULL,
  cutoff = (function(x) {
     x <- -20
 }),
  ...
)
}
\arguments{
\item{barplot.data}{A dataframe with Columns Sum, of scores, and Sample}

\item{filename}{A filename where to save plot. If NULL the barplot object will be returned}

\item{cutoff}{A function to plot the cutoff fot what is deemed a poor sample}
}
\value{
the barplot or NULL depending if filename is specified
}
\description{
This function takes the barplot data generated by organise.barplot.data
and returns the barplot if no filename is specified. If filename is
is specified it saves the plot as file and returns NULL.
}
