% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.standard.plots.R
\name{get.qc.heatmap}
\alias{get.qc.heatmap}
\title{Generates the standard heatmap of scores for each sample}
\usage{
get.qc.heatmap(qc.data.heatmap, ylabels, ...)
}
\arguments{
\item{qc.data.heatmap}{A dataframe of sign-corrected scores for each Sample}

\item{filename}{A filename to save heatmap. If NULL the heatmap object will be returned}
}
\value{
the heatmap or NULL depending if filename is specified
}
\description{
This function takes the the scores for each sample and each metric,
after being sign-corrected, and returns the standard heatmap, if filename
is NULL. If filename is not NULL it saves the heatmap to file and returns NULL
}
