% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.standard.plots.R
\name{get.qc.multipanelplot}
\alias{get.qc.multipanelplot}
\title{Generates the multipanel plot of heatmap and barplot}
\usage{
get.qc.multipanelplot(barplot, heatmap, filename, ...)
}
\arguments{
\item{barplot}{A barplot of the samples aggregated score}

\item{heatmap}{A heatmap of the sign-corrected scores for each sample}

\item{filename}{A filename to save the multipanel plot}
}
\description{
This function takes the barplot and heatmap and returns the multipanel plot
of the two.
}
