% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ylabels}
\alias{ylabels}
\title{Formatted QC metrics labels}
\format{
A character vector of formatted QC metric labels
}
\usage{
data(ylabels)
}
\description{
Formatted QC metrics labels
}
\examples{
\donttest{
data(ylabels)
data(example.qc.dataframe)
data(sign.correction)

zscores <- zscores.from.metrics(qc.data = example.qc.dataframe);
zscores.corrected <- correct.zscore.signs(
  zscores = zscores,
  signs.data = sign.correction,
  metric.col.name = 'Metric',
  signs.col.name = 'Sign'
);
quality.scores <- accumulate.zscores(zscores.corrected = zscores.corrected);

qc.heatmap <- get.qc.heatmap(
  zscores = zscores.corrected,
  quality.scores = quality.scores,
  yaxis.lab = ylabels
);
}
}
\keyword{datasets}
