% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.zscore.R
\name{zscores.from.metrics}
\alias{zscores.from.metrics}
\title{Calculate z-scores for each metric across each sample}
\usage{
zscores.from.metrics(qc.data, filename = NULL)
}
\arguments{
\item{qc.data}{A dataframe whose rows are samples and each column a QC metric}

\item{filename}{A filename where to save data. If NULL data will not be saved to file}
}
\value{
A dataframe of z-scores for each metric
}
\description{
This function takes a dataframe of QC metrics, and calculcates the
the z-scores. If filename is specified, the results will be save to file.
}
