% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate.zscores.R
\name{accumulate.zscores}
\alias{accumulate.zscores}
\title{Sum across sign corrected z-scores for total sample quality score}
\usage{
accumulate.zscores(zscores.corrected, filename = NULL)
}
\arguments{
\item{zscores.corrected}{A dataframe whose rows are samples and each column a QC metric}

\item{filename}{A filename where to save data. If NULL data will not be saved to file}
}
\value{
A dataframe of aggregated z-scores for each sample
\describe{
\item{Sample}{Sample IDs defined by the rownames of \code{zscores.corrected}}
\item{Sum}{Sum of z-scores}
}
}
\description{
This function takes a dataframe of all the sign corrected scores, thus all negative, and aggregates
to get a total sample quality score.
}
