% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.zscore.R
\name{accumulate.zscores}
\alias{accumulate.zscores}
\title{Sum across sign corrected z-scores for total sample score}
\usage{
accumulate.zscores(zscores.corrected, filename = NULL)
}
\arguments{
\item{zscores.corrected}{A dataframe whose rows are samples and each column a QC metric}

\item{filename}{A filename where to save data. If NULL data will not be saved to file}
}
\value{
A dataframe of z-scores for each metric
}
\description{
This function takes a dataframe of all the negative scores and aggregates
to a form which can be plotted as a a barplot.
}
